requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.9
Django==3.2.3
django-annoying==0.10.6
django-extensions==3.1.3
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
importlib-metadata==4.3.0
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
zipp==3.4.1
